move process_raw_raster to download.py
